update the styl of dropdown in header navbar
