--- src/components/header/v2/constants.js
@@ -1,156 +1,0 @@
-export const LANGUAGES = [
-  {
-    label: '中文',
-    value: 'cn',
-  },
-  {
-    label: 'English',
-    value: 'en',
-  },
-];
-
-export const NAVLIST_EN = [
-  {
-    label: 'Milvus 2021 survey!',
-    link: 'https://milvus.typeform.com/to/EgMCxy2T',
-    activeKey: null,
-    className: 'fancyLink',
-  },
-  {
-    label: 'Docs',
-    link: '/docs',
-    activeKey: '/docs',
-  },
-  {
-    label: 'Blog',
-    link: '/blog',
-    activeKey: '/blog',
-  },
-  {
-    label: 'Forum',
-    link: 'https://discuss.milvus.io/',
-    activeKey: 'null',
-  },
-  {
-    label: 'Tutorials',
-    link: null,
-    activeKey: null,
-    subMenu: [
-      {
-        label: 'Bootcamp',
-        link: '/bootcamp',
-        activeKey: 'bootcamp',
-      },
-      {
-        label: 'Live Demos',
-        link: 'https://zilliz.com/milvus-demos?isZilliz=true',
-        activeKey: null,
-      },
-      {
-        label: 'Video',
-        link: 'https://www.youtube.com/channel/UCMCo_F7pKjMHBlfyxwOPw-g',
-        activeKey: null,
-      },
-    ],
-  },
-  {
-    label: 'Tools',
-    link: null,
-    activeKey: null,
-    subMenu: [
-      {
-        label: 'Sizing Tool',
-        link: 'https://zilliz.com/sizing-tool',
-        activeKey: null,
-      },
-      {
-        label: 'Insight',
-        link: 'https://github.com/milvus-io/milvus-insight',
-        activeKey: null,
-      },
-      {
-        label: 'Milvus CLI',
-        link: 'https://github.com/milvus-io/milvus_cli',
-        activeKey: null,
-      },
-    ],
-  },
-  {
-    label: 'Community',
-    link: '/community',
-    activeKey: '/community',
-  },
-];
-
-export const NAVLIST_CN = [
-  {
-    label: 'Milvus 2021 问卷!',
-    link: 'https://milvus.typeform.com/to/GM5f8HOe',
-    activeKey: null,
-    className: 'fancyLink',
-  },
-  {
-    label: '文档',
-    link: '/docs',
-    activeKey: '/docs',
-  },
-  {
-    label: '博客',
-    link: '/blog',
-    activeKey: '/blog',
-  },
-  {
-    label: '论坛',
-    link: 'https://discuss.milvus.io/',
-    activeKey: 'null',
-  },
-  {
-    label: '教程',
-    link: null,
-    activeKey: null,
-    subMenu: [
-      {
-        label: '训练营',
-        link: '/bootcamp',
-        activeKey: 'bootcamp',
-      },
-      {
-        label: '在线 Demos',
-        link: 'https://zilliz.com/milvus-demos?isZilliz=true',
-        activeKey: null,
-      },
-      {
-        label: '视频',
-        link: 'https://www.youtube.com/channel/UCMCo_F7pKjMHBlfyxwOPw-g',
-        activeKey: null,
-      },
-    ],
-  },
-  {
-    label: '工具',
-    link: null,
-    activeKey: null,
-    subMenu: [
-      {
-        label: 'Sizing Tool',
-        link: 'https://zilliz.com/sizing-tool',
-        activeKey: null,
-      },
-      {
-        label: 'Insight',
-        link: 'https://github.com/milvus-io/milvus-insight',
-        activeKey: null,
-      },
-      {
-        label: 'Milvus CLI',
-        link: 'https://github.com/milvus-io/milvus_cli',
-        activeKey: null,
-      },
-    ],
-  },
-  {
-    label: '社区',
-    link: '/community',
-    activeKey: '/community',
-  },
-];
