+++ src/components/header/v2/constants.jsx
@@ -1,0 +1,280 @@
+import React from 'react';
+
+const Bootcamp = () => (
+  <svg
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M20.6235 15.9401L16.5512 12.0939L20.6229 8.2484C20.7424 8.13568 20.8372 8.00183 20.9019 7.8545C20.9666 7.70717 20.9999 7.54924 20.9999 7.38975C21 7.23025 20.9668 7.0723 20.9022 6.92493C20.8376 6.77755 20.7429 6.64364 20.6235 6.53084L20.6229 6.52963L18.3216 4.35612C18.2022 4.24328 18.0605 4.15375 17.9045 4.09264C17.7485 4.03154 17.5813 4.00006 17.4124 4C17.2435 3.99994 17.0763 4.03131 16.9203 4.09231C16.7642 4.15331 16.6224 4.24275 16.503 4.35552L16.5024 4.35612L12.4307 8.20165L8.35839 4.35552C8.11696 4.12847 7.79013 4.001 7.44943 4.001C7.10873 4.001 6.7819 4.12847 6.54047 4.35552L3.37648 7.34317C3.13542 7.57098 3 7.87988 3 8.20195C3 8.52403 3.13542 8.83292 3.37648 9.06073L7.44814 12.9063L3.00042 17.1082V20.9998H7.12094L11.5693 16.7985L15.6416 20.6459C15.8834 20.8727 16.2104 21 16.5512 21C16.892 21 17.219 20.8727 17.4608 20.6459L20.6235 17.6576C20.8646 17.4298 21 17.1209 21 16.7988C21 16.4768 20.8646 16.1679 20.6235 15.9401ZM17.4081 5.21217L19.7159 7.39174L15.6448 11.2373L13.3371 9.05769L17.4081 5.21217ZM6.85739 19.7856H4.28608V17.6054L8.35517 13.7629L10.7182 15.9959L6.85739 19.7856ZM16.5512 19.788L4.28608 8.20165L7.45007 5.2146L9.8851 7.51378L8.43745 8.88163L9.34705 9.7401L10.7941 8.37286L13.4489 10.8803L12 12.2463L12.9102 13.1048L14.3579 11.7375L17.0128 14.245L15.5651 15.6116L16.4741 16.4707L17.9217 15.104L19.7152 16.7985L16.5512 19.788Z"
+      fill="#000"
+    />
+  </svg>
+);
+
+const Cli = () => (
+  <svg
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M7.84845 18.0886C7.68809 18.0854 7.53421 18.0248 7.41474 17.9178L2.22344 13.239C2.15283 13.1793 2.09622 13.1048 2.05761 13.0207C2.019 12.9367 1.99934 12.8452 2.00002 12.7527C2.01084 12.5684 2.09059 12.3948 2.22344 12.2665L7.48046 7.69286C7.54443 7.63141 7.62019 7.58355 7.70316 7.55216C7.78612 7.52077 7.87459 7.50651 7.96322 7.51023C8.05185 7.51395 8.1388 7.53557 8.21885 7.5738C8.2989 7.61203 8.37038 7.66607 8.42898 7.73266C8.48758 7.79926 8.53209 7.87702 8.55983 7.96128C8.58757 8.04554 8.59797 8.13454 8.59039 8.22292C8.58281 8.31131 8.55741 8.39724 8.51573 8.47554C8.47405 8.55385 8.41695 8.6229 8.34786 8.67855L3.65598 12.7527L8.29529 16.9715C8.35689 17.0326 8.40577 17.1053 8.43913 17.1853C8.4725 17.2654 8.48967 17.3513 8.48967 17.4381C8.48967 17.5248 8.4725 17.6107 8.43913 17.6908C8.40577 17.7708 8.35689 17.8435 8.29529 17.9046C8.17512 18.0205 8.01536 18.0863 7.84845 18.0886ZM16.9299 18.0886C16.8379 18.0868 16.7472 18.066 16.6635 18.0275C16.5799 17.9891 16.505 17.9338 16.4437 17.8652C16.3224 17.7401 16.2546 17.5728 16.2546 17.3986C16.2546 17.2244 16.3224 17.0571 16.4437 16.9321L21.083 12.7527L16.4962 8.67855C16.4346 8.61746 16.3858 8.54478 16.3524 8.4647C16.319 8.38462 16.3019 8.29873 16.3019 8.21199C16.3019 8.12524 16.319 8.03935 16.3524 7.95927C16.3858 7.87919 16.4346 7.80651 16.4962 7.74543C16.5562 7.68423 16.6278 7.63561 16.7068 7.60242C16.7857 7.56923 16.8706 7.55214 16.9562 7.55214C17.0419 7.55214 17.1267 7.56923 17.2057 7.60242C17.2847 7.63561 17.3563 7.68423 17.4162 7.74543L22.6732 12.319C22.7429 12.3795 22.7987 12.4542 22.837 12.5381C22.8752 12.622 22.895 12.7131 22.895 12.8053C22.895 12.8975 22.8752 12.9886 22.837 13.0725C22.7987 13.1564 22.7429 13.2311 22.6732 13.2916L17.4951 17.9703C17.4141 18.0278 17.3221 18.068 17.2248 18.0883C17.1276 18.1087 17.0272 18.1088 16.9299 18.0886ZM10.9764 20.3097H10.8712C10.7101 20.278 10.5661 20.1882 10.4668 20.0574C10.3675 19.9265 10.3196 19.7638 10.3324 19.6L12.6455 5.49805C12.6776 5.32677 12.7751 5.17471 12.9174 5.07412C13.0597 4.97353 13.2356 4.93231 13.4077 4.95921C13.4928 4.97193 13.5745 5.00157 13.6479 5.04637C13.7213 5.09118 13.785 5.15023 13.8352 5.22006C13.8855 5.28989 13.9212 5.36907 13.9403 5.45293C13.9594 5.53679 13.9616 5.62363 13.9466 5.70833L11.6335 19.7709C11.6055 19.9239 11.5241 20.0621 11.4038 20.1608C11.2834 20.2595 11.132 20.3122 10.9764 20.3097Z"
+      fill="black"
+    />
+  </svg>
+);
+
+const Demos = () => (
+  <svg
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M19.3999 3.59961H4.60008C3.7164 3.59961 3 4.31601 3 5.19969V18.7995C3 19.6831 3.7164 20.3996 4.60008 20.3996H19.3999C20.2837 20.3996 21 19.6831 21 18.7995V5.19969C21 4.31601 20.2837 3.59961 19.3999 3.59961ZM18.4 5.59953C18.5566 5.60337 18.7055 5.66827 18.8149 5.78038C18.9243 5.8925 18.9855 6.04294 18.9855 6.19959C18.9855 6.35624 18.9243 6.50668 18.8149 6.61879C18.7055 6.73091 18.5566 6.79581 18.4 6.79965C18.2408 6.79962 18.0882 6.7364 17.9757 6.62389C17.8632 6.51137 17.7999 6.35878 17.7998 6.19965C17.7999 6.04052 17.8632 5.88793 17.9757 5.77541C18.0882 5.66288 18.2408 5.59962 18.4 5.59953ZM15.9998 5.59953C16.1564 5.60337 16.3053 5.66827 16.4147 5.78038C16.5241 5.8925 16.5854 6.04294 16.5854 6.19959C16.5854 6.35624 16.5241 6.50668 16.4147 6.61879C16.3053 6.73091 16.1564 6.79581 15.9998 6.79965C15.8422 6.79742 15.6917 6.73322 15.581 6.62093C15.4703 6.50864 15.4082 6.35728 15.4082 6.19959C15.4082 6.0419 15.4703 5.89054 15.581 5.77825C15.6917 5.66596 15.8422 5.60176 15.9998 5.59953ZM13.6001 5.59953C13.7578 5.60176 13.9082 5.66596 14.0189 5.77825C14.1296 5.89054 14.1917 6.0419 14.1917 6.19959C14.1917 6.35728 14.1296 6.50864 14.0189 6.62093C13.9082 6.73322 13.7578 6.79742 13.6001 6.79965C13.4409 6.79965 13.2883 6.73643 13.1758 6.6239C13.0632 6.51136 13 6.35874 13 6.19959C13 6.04044 13.0632 5.88782 13.1758 5.77528C13.2883 5.66275 13.4409 5.59953 13.6001 5.59953ZM19.8001 17.9996C19.8001 18.8829 19.4837 19.1994 18.5999 19.1994H5.4C4.51644 19.1994 4.2 18.8829 4.2 17.9996V7.99965H19.8V17.9996H19.8001ZM6.4638 12.7076L8.72292 13.3131L6.59208 14.5433C6.50024 14.5963 6.43322 14.6837 6.40577 14.7861C6.37832 14.8886 6.39267 14.9977 6.44568 15.0896L6.6456 15.4363C6.67183 15.4818 6.7068 15.5217 6.74848 15.5537C6.79017 15.5857 6.83776 15.6092 6.88853 15.6227C6.9393 15.6363 6.99224 15.6397 7.04433 15.6328C7.09642 15.6259 7.14663 15.6087 7.19208 15.5823L10.3667 13.7496C10.4644 13.7659 10.5646 13.7454 10.648 13.6919C10.7313 13.6384 10.7918 13.5558 10.8175 13.4601L10.9211 13.0738C10.9484 12.9714 10.934 12.8622 10.8809 12.7704C10.8279 12.6786 10.7406 12.6115 10.6381 12.584L10.4153 12.5243C10.379 12.5085 10.3406 12.4982 10.3013 12.4938L6.77448 11.5488C6.72375 11.5351 6.67084 11.5316 6.61876 11.5385C6.56667 11.5453 6.51645 11.5623 6.47095 11.5886C6.42545 11.6148 6.38556 11.6498 6.35358 11.6914C6.32159 11.7331 6.29812 11.7806 6.28452 11.8314L6.18108 12.2179C6.16747 12.2686 6.16399 12.3215 6.17083 12.3736C6.17768 12.4256 6.19471 12.4759 6.22097 12.5213C6.24723 12.5668 6.28218 12.6067 6.32385 12.6386C6.36552 12.6706 6.41307 12.694 6.4638 12.7076ZM13.4002 16.7996H17.4C17.4525 16.7996 17.5046 16.7893 17.5531 16.7692C17.6016 16.7491 17.6457 16.7196 17.6828 16.6825C17.72 16.6453 17.7494 16.6012 17.7695 16.5526C17.7896 16.5041 17.7999 16.4521 17.7998 16.3995V15.9997C17.7999 15.9471 17.7896 15.8951 17.7695 15.8465C17.7494 15.798 17.72 15.7539 17.6829 15.7167C17.6457 15.6795 17.6016 15.65 17.5531 15.6299C17.5046 15.6098 17.4525 15.5994 17.4 15.5994H13.4002C13.3476 15.5994 13.2955 15.6097 13.247 15.6298C13.1984 15.6499 13.1543 15.6794 13.1171 15.7166C13.0799 15.7538 13.0505 15.7979 13.0304 15.8465C13.0103 15.8951 12.9999 15.9471 13 15.9997V16.3995C12.9999 16.4521 13.0103 16.5041 13.0304 16.5527C13.0505 16.6013 13.0799 16.6454 13.1171 16.6825C13.1543 16.7197 13.1984 16.7492 13.247 16.7693C13.2955 16.7893 13.3476 16.7997 13.4002 16.7996Z"
+      fill="black"
+    />
+  </svg>
+);
+
+const Insight = () => (
+  <svg
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M8.25 21C8.05109 21 7.86032 21.079 7.71967 21.2197C7.57902 21.3603 7.5 21.5511 7.5 21.75C7.5 21.9489 7.57902 22.1397 7.71967 22.2803C7.86032 22.421 8.05109 22.5 8.25 22.5V21ZM15.75 22.5C15.9489 22.5 16.1397 22.421 16.2803 22.2803C16.421 22.1397 16.5 21.9489 16.5 21.75C16.5 21.5511 16.421 21.3603 16.2803 21.2197C16.1397 21.079 15.9489 21 15.75 21V22.5ZM15.8565 18.0105L15.1327 17.814L15.8565 18.0105ZM8.1435 18.0105L8.86725 17.814L8.1435 18.0105ZM15 12.75C15 12.5511 14.921 12.3603 14.7803 12.2197C14.6397 12.079 14.4489 12 14.25 12C14.0511 12 13.8603 12.079 13.7197 12.2197C13.579 12.3603 13.5 12.5511 13.5 12.75H15ZM10.5 12.75C10.5 12.5511 10.421 12.3603 10.2803 12.2197C10.1397 12.079 9.94891 12 9.75 12C9.55109 12 9.36032 12.079 9.21967 12.2197C9.07902 12.3603 9 12.5511 9 12.75H10.5ZM8.25 22.5H15.75V21H8.25V22.5ZM5.25 9.5835C5.25 5.58075 8.2275 3 12 3V1.5C7.488 1.5 3.75 4.6665 3.75 9.5835H5.25ZM12 3C15.7725 3 18.75 5.58075 18.75 9.5835H20.25C20.25 4.6665 16.512 1.5 12 1.5V3ZM18.75 9.5835C18.75 11.097 18.0855 12.4905 17.2635 13.8622C17.0602 14.2028 16.851 14.5342 16.6425 14.8657C16.4363 15.1942 16.2293 15.5243 16.038 15.8468C15.66 16.4843 15.3127 17.1488 15.1327 17.814L16.5802 18.2062C16.7055 17.745 16.9657 17.2238 17.3287 16.611C17.508 16.308 17.7037 15.996 17.9122 15.6645C18.1192 15.336 18.3367 14.9895 18.5505 14.634C19.398 13.2195 20.25 11.5275 20.25 9.5835H18.75ZM8.86725 17.814C8.68725 17.1488 8.34 16.485 7.962 15.8468C7.77075 15.5243 7.5645 15.1942 7.3575 14.8657C7.149 14.5342 6.93975 14.2028 6.7365 13.8622C5.9145 12.4905 5.25 11.097 5.25 9.5835H3.75C3.75 11.5275 4.602 13.2195 5.4495 14.634C5.66325 14.9902 5.88075 15.336 6.087 15.6645C6.29625 15.996 6.492 16.308 6.672 16.611C7.03425 17.2238 7.2945 17.745 7.41975 18.2062L8.86725 17.814ZM15 18H9V19.5H15V18ZM7.41975 18.2062C7.58775 18.8258 8.14725 19.5 9 19.5V18L8.9955 17.9985C8.98275 17.9925 8.97109 17.9844 8.961 17.9745C8.91566 17.9306 8.88323 17.8751 8.86725 17.814L7.41975 18.2062ZM15.1327 17.814C15.1168 17.8751 15.0843 17.9306 15.039 17.9745C15.0277 17.9853 15.0145 17.994 15 18V19.5C15.8527 19.5 16.4123 18.8258 16.5802 18.2062L15.1327 17.814ZM13.5 12.75C13.5 13.1478 13.342 13.5294 13.0607 13.8107C12.7794 14.092 12.3978 14.25 12 14.25V15.75C12.7956 15.75 13.5587 15.4339 14.1213 14.8713C14.6839 14.3087 15 13.5456 15 12.75H13.5ZM12 14.25C11.6022 14.25 11.2206 14.092 10.9393 13.8107C10.658 13.5294 10.5 13.1478 10.5 12.75H9C9 13.5456 9.31607 14.3087 9.87868 14.8713C10.4413 15.4339 11.2044 15.75 12 15.75V14.25Z"
+      fill="#000"
+    />
+  </svg>
+);
+
+const Tool = () => (
+  <svg
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M15.75 8.25H19.5C19.6989 8.25 19.8897 8.32902 20.0303 8.46967C20.171 8.61032 20.25 8.80109 20.25 9C20.25 9.19891 20.171 9.38968 20.0303 9.53033C19.8897 9.67098 19.6989 9.75 19.5 9.75H15C14.8011 9.75 14.6103 9.67098 14.4697 9.53033C14.329 9.38968 14.25 9.19891 14.25 9V4.5C14.25 4.30109 14.329 4.11032 14.4697 3.96967C14.6103 3.82902 14.8011 3.75 15 3.75C15.1989 3.75 15.3897 3.82902 15.5303 3.96967C15.671 4.11032 15.75 4.30109 15.75 4.5V8.25Z"
+      fill="black"
+    />
+    <path
+      d="M15.5325 9.53264C15.3913 9.67387 15.1997 9.75321 15 9.75321C14.8003 9.75321 14.6087 9.67387 14.4675 9.53264C14.3263 9.39141 14.2469 9.19986 14.2469 9.00014C14.2469 8.80041 14.3263 8.60887 14.4675 8.46764L19.7175 3.21764C19.8587 3.07641 20.0503 2.99707 20.25 2.99707C20.4497 2.99707 20.6413 3.07641 20.7825 3.21764C20.9237 3.35887 21.0031 3.55041 21.0031 3.75014C21.0031 3.94987 20.9237 4.14141 20.7825 4.28264L15.5325 9.53264ZM8.25 8.25014V4.50014C8.25 4.30123 8.32902 4.11046 8.46967 3.96981C8.61032 3.82916 8.80109 3.75014 9 3.75014C9.19891 3.75014 9.38968 3.82916 9.53033 3.96981C9.67098 4.11046 9.75 4.30123 9.75 4.50014V9.00014C9.75 9.19905 9.67098 9.38982 9.53033 9.53047C9.38968 9.67112 9.19891 9.75014 9 9.75014H4.5C4.30109 9.75014 4.11032 9.67112 3.96967 9.53047C3.82902 9.38982 3.75 9.19905 3.75 9.00014C3.75 8.80123 3.82902 8.61046 3.96967 8.46981C4.11032 8.32916 4.30109 8.25014 4.5 8.25014H8.25Z"
+      fill="black"
+    />
+    <path
+      d="M9.53264 8.46764C9.67387 8.60887 9.75321 8.80041 9.75321 9.00014C9.75321 9.19986 9.67387 9.39141 9.53264 9.53264C9.39141 9.67387 9.19986 9.75321 9.00014 9.75321C8.80041 9.75321 8.60887 9.67387 8.46764 9.53264L3.21764 4.28264C3.07641 4.14141 2.99707 3.94987 2.99707 3.75014C2.99707 3.55041 3.07641 3.35887 3.21764 3.21764C3.35887 3.07641 3.55041 2.99707 3.75014 2.99707C3.94987 2.99707 4.14141 3.07641 4.28264 3.21764L9.53264 8.46764ZM15.7501 15.7501V19.5001C15.7501 19.699 15.6711 19.8898 15.5305 20.0305C15.3898 20.1711 15.199 20.2501 15.0001 20.2501C14.8012 20.2501 14.6105 20.1711 14.4698 20.0305C14.3292 19.8898 14.2501 19.699 14.2501 19.5001V15.0001C14.2501 14.8012 14.3292 14.6105 14.4698 14.4698C14.6105 14.3292 14.8012 14.2501 15.0001 14.2501H19.5001C19.699 14.2501 19.8898 14.3292 20.0305 14.4698C20.1711 14.6105 20.2501 14.8012 20.2501 15.0001C20.2501 15.199 20.1711 15.3898 20.0305 15.5305C19.8898 15.6711 19.699 15.7501 19.5001 15.7501H15.7501Z"
+      fill="black"
+    />
+    <path
+      d="M14.4675 15.5326C14.3976 15.4627 14.3421 15.3797 14.3043 15.2883C14.2664 15.197 14.2469 15.099 14.2469 15.0001C14.2469 14.9012 14.2664 14.8033 14.3043 14.712C14.3421 14.6206 14.3976 14.5376 14.4675 14.4676C14.5374 14.3977 14.6204 14.3422 14.7118 14.3044C14.8032 14.2665 14.9011 14.2471 15 14.2471C15.0989 14.2471 15.1968 14.2665 15.2882 14.3044C15.3796 14.3422 15.4626 14.3977 15.5325 14.4676L20.7825 19.7176C20.9237 19.8589 21.0031 20.0504 21.0031 20.2501C21.0031 20.4499 20.9237 20.6414 20.7825 20.7826C20.6413 20.9239 20.4497 21.0032 20.25 21.0032C20.0503 21.0032 19.8587 20.9239 19.7175 20.7826L14.4675 15.5326ZM4.5 15.7501C4.30109 15.7501 4.11032 15.6711 3.96967 15.5305C3.82902 15.3898 3.75 15.1991 3.75 15.0001C3.75 14.8012 3.82902 14.6105 3.96967 14.4698C4.11032 14.3292 4.30109 14.2501 4.5 14.2501H9C9.19891 14.2501 9.38968 14.3292 9.53033 14.4698C9.67098 14.6105 9.75 14.8012 9.75 15.0001V19.5001C9.75 19.6991 9.67098 19.8898 9.53033 20.0305C9.38968 20.1711 9.19891 20.2501 9 20.2501C8.80109 20.2501 8.61032 20.1711 8.46967 20.0305C8.32902 19.8898 8.25 19.6991 8.25 19.5001V15.7501H4.5Z"
+      fill="black"
+    />
+    <path
+      d="M8.46764 14.4676C8.53757 14.3977 8.62059 14.3422 8.71195 14.3044C8.80332 14.2665 8.90124 14.2471 9.00014 14.2471C9.09903 14.2471 9.19696 14.2665 9.28833 14.3044C9.37969 14.3422 9.46271 14.3977 9.53264 14.4676C9.60257 14.5376 9.65804 14.6206 9.69588 14.712C9.73373 14.8033 9.75321 14.9012 9.75321 15.0001C9.75321 15.099 9.73373 15.197 9.69588 15.2883C9.65804 15.3797 9.60257 15.4627 9.53264 15.5326L4.28264 20.7826C4.14141 20.9239 3.94987 21.0032 3.75014 21.0032C3.55041 21.0032 3.35887 20.9239 3.21764 20.7826C3.07641 20.6414 2.99707 20.4499 2.99707 20.2501C2.99707 20.0504 3.07641 19.8589 3.21764 19.7176L8.46764 14.4676Z"
+      fill="black"
+    />
+  </svg>
+);
+
+const Video = () => (
+  <svg
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M20.5292 3.68652H3.41982C2.37217 3.68652 1.51904 4.53965 1.51904 5.5873V18.2951C1.51904 19.3428 2.37217 20.1959 3.41982 20.1959H20.5292C21.5769 20.1959 22.43 19.3428 22.43 18.2951V5.5873C22.43 4.53965 21.5769 3.68652 20.5292 3.68652ZM20.555 18.2951C20.555 18.3092 20.5433 18.3209 20.5292 18.3209H3.41982C3.40576 18.3209 3.39404 18.3092 3.39404 18.2951V5.5873C3.39404 5.57324 3.40576 5.56152 3.41982 5.56152H20.5292C20.5433 5.56152 20.555 5.57324 20.555 5.5873V18.2951Z"
+      fill="black"
+    />
+    <path
+      d="M15.1708 11.126L9.9958 8.18692C9.70518 8.02286 9.34893 8.02286 9.06064 8.19161C8.77236 8.36036 8.59424 8.66739 8.59424 9.00255V14.8854C8.59424 15.2182 8.77236 15.5275 9.06064 15.6963C9.20596 15.7807 9.37002 15.8229 9.53174 15.8229C9.69111 15.8229 9.85049 15.783 9.9958 15.701L15.1708 12.7596C15.4638 12.5932 15.6442 12.2815 15.6442 11.944C15.6442 11.6065 15.4638 11.2947 15.1708 11.126ZM10.4692 13.2729V10.6127L12.8106 11.9416L10.4692 13.2729Z"
+      fill="black"
+    />
+  </svg>
+);
+
+export const LANGUAGES = [
+  {
+    label: '中文',
+    value: 'cn',
+  },
+  {
+    label: 'English',
+    value: 'en',
+  },
+];
+
+export const NAVLIST_EN = [
+  {
+    label: 'Milvus 2021 survey!',
+    link: 'https://milvus.typeform.com/to/EgMCxy2T',
+    activeKey: null,
+    className: 'fancyLink',
+  },
+  {
+    label: 'Docs',
+    link: '/docs',
+    activeKey: '/docs',
+  },
+  {
+    label: 'Blog',
+    link: '/blog',
+    activeKey: '/blog',
+  },
+  {
+    label: 'Forum',
+    link: 'https://discuss.milvus.io/',
+    activeKey: 'null',
+  },
+  {
+    label: 'Tutorials',
+    link: null,
+    activeKey: null,
+    subMenu: [
+      {
+        label: 'Bootcamp',
+        link: '/bootcamp',
+        activeKey: 'bootcamp',
+        icon: Bootcamp,
+      },
+      {
+        label: 'Live Demos',
+        link: 'https://zilliz.com/milvus-demos?isZilliz=true',
+        activeKey: null,
+        icon: Demos,
+      },
+      {
+        label: 'Video',
+        link: 'https://www.youtube.com/channel/UCMCo_F7pKjMHBlfyxwOPw-g',
+        activeKey: null,
+        icon: Video,
+      },
+    ],
+  },
+  {
+    label: 'Tools',
+    link: null,
+    activeKey: null,
+    subMenu: [
+      {
+        label: 'Sizing Tool',
+        link: 'https://zilliz.com/sizing-tool',
+        activeKey: null,
+        icon: Tool,
+      },
+      {
+        label: 'Insight',
+        link: 'https://github.com/milvus-io/milvus-insight',
+        activeKey: null,
+        icon: Insight,
+      },
+      {
+        label: 'Milvus CLI',
+        link: 'https://github.com/milvus-io/milvus_cli',
+        activeKey: null,
+        icon: Cli,
+      },
+    ],
+  },
+  {
+    label: 'Community',
+    link: '/community',
+    activeKey: '/community',
+  },
+];
+
+export const NAVLIST_CN = [
+  {
+    label: 'Milvus 2021 问卷!',
+    link: 'https://milvus.typeform.com/to/GM5f8HOe',
+    activeKey: null,
+    className: 'fancyLink',
+  },
+  {
+    label: '文档',
+    link: '/docs',
+    activeKey: '/docs',
+  },
+  {
+    label: '博客',
+    link: '/blog',
+    activeKey: '/blog',
+  },
+  {
+    label: '论坛',
+    link: 'https://discuss.milvus.io/',
+    activeKey: 'null',
+  },
+  {
+    label: '教程',
+    link: null,
+    activeKey: null,
+    subMenu: [
+      {
+        label: '训练营',
+        link: '/bootcamp',
+        activeKey: 'bootcamp',
+        icon: Bootcamp,
+      },
+      {
+        label: '在线 Demos',
+        link: 'https://zilliz.com/milvus-demos?isZilliz=true',
+        activeKey: null,
+        icon: Demos,
+      },
+      {
+        label: '视频',
+        link: 'https://www.youtube.com/channel/UCMCo_F7pKjMHBlfyxwOPw-g',
+        activeKey: null,
+        icon: Video,
+      },
+    ],
+  },
+  {
+    label: '工具',
+    link: null,
+    activeKey: null,
+    subMenu: [
+      {
+        label: 'Sizing Tool',
+        link: 'https://zilliz.com/sizing-tool',
+        activeKey: null,
+        icon: Tool,
+      },
+      {
+        label: 'Insight',
+        link: 'https://github.com/milvus-io/milvus-insight',
+        activeKey: null,
+        icon: Insight,
+      },
+      {
+        label: 'Milvus CLI',
+        link: 'https://github.com/milvus-io/milvus_cli',
+        activeKey: null,
+        icon: Cli,
+      },
+    ],
+  },
+  {
+    label: '社区',
+    link: '/community',
+    activeKey: '/community',
+  },
+];
